Add partial implementation of UserReportService
The UserReportService is designed to return csvs of consumers (within
the backend consumers are users). Right now it returns a string that
contains the csv data.

In order to do this work I needed to make a UserService to encapsulate
the retrieval of users in order to do the reporting. At this time the
User Service is NOT hooked up to the database and uses mocked data
everywhere. The next commit should introduce the database connection. At
that point we can begin using the UserService and UserReportService.