Query users properly and filter by api
This should add all functionality needed to query users/consumers and
generate the csv report. There a few caveats.

1. The dynamoService had preset parameters that could be passed in to
use the AWS DynamoDB querying. I didn't want to go through and refactor
a ton of code at this time, so I created a second method, called
hardScan(), that takes the raw AWS ScanInput type. This lets us query
more dynamically.

2. Updated the UserService and UserReportService to set the user
oAuthApplicationType fields to blank. We don't store that data in
dynamo, and so when we retrieve them those fields don't exist. In order
to get the test to pass easily we just set them to blank. The field is
needed to use the User constructor so I also couldn't just leave it
null or undefined.

3. The table name in UserService was wrong, or at least not the same as
what's in the User model file. This should now be fixed.

4. The api filtering is done within our node server in getUser()
function. Apis are stored as a comma delimited list in dynamo at this
time, so querying by the appropriate APIs would be complicated. This
should work for the moment, and if we have performance issues we can
investigate writing a dynamo query or restructuring the dynamo table to
store the  fields as an actual array, which should resolve some of the
complication.

5. I applied some lint fixes, notable in UserService. I moved a function
above the class definition for UserService.