Add APIs to csv and support for fields with commas
I added APIs to the csv because it seemed like useful information, even
if just to confirm that the file open looks like it has the right
consumers on it.

Because APIs are stored as a comma separated list in dynamo and stored
that way in the User object currently, I encapsulated each field in
double quotes. This should cause excel and other spreadsheet tools to
open the csv properly. I tested with OpenOffice.