Use string type for channels and more err handling
When sending file uploads to slack, the channels field should be a comma
separated string, not an array of strings.

I also added some basic error handling for logging if the file upload
doesn't work. This already came in handy because the slack token didn't
have the permission to write files.