Add validation to csv consumer reports endpoint
The consumer reports endpoint now checks for valid APIs in the given
list, the same way that the developer_application endpoint takes them.

I moved the custom validation function out of the DeveloperApplication
handler and into a separate file [made a schemaValidators folder for
now]. I then imported this function into the UserReports handler for use
in the validation.