Remove duplicate users from the csv report
I added a removeDuplicateUsers() function to the UserService. At some
point we'll be merging duplicate users in the database, when we add
accounts, and will likely reuse this functionality, so I decided to
locate it there.

At the moment this is only used while generating the consumer reports in
the UserReportService, but again, we'll likely use it when merging and
duplicating accounts, whether on retrieval, user applications, or
somewhere else.