Use consistent DynamoDB package types
For submitting data to dynamoDB, we are using the DocumentClient. This
is confusing because within DocumentClient (a namespace), there are
multiple convenience types that are different than if you import those
same types from the straight aws-sdk dynamodb package.

For instance, ScanInput accepts different values than
DocumentClient.ScanInput.

This caused errors while I was trying to write a dynamo db query because
the IDEs and typescript compilation doesn't detect that the types don't
match. This caused the queries to 'succeed' but not return any results,
because the mismatch was causing dynamo to search for string matches to
odd data types.

This change doesn't include any functional changes. It is purely to
fix the incorrect types throughout the application and any resulting
modifications due to the type updates.

As a note, I had to import DocumentClient, and then reference it every
place I needed one of these hidden APIs [DocumentClient.ScanInput for
example]. I could not find a way with typescript to easily import the
types - I think because there seems to be both a DocumentClient class
and a DocumentClient namespace in the package, which I think causes
other conflicts. For now this should be adequate for us to continue
working, but if we continue using dynamo, we should try and get all of
this sorted out.