Fix broken SlackService tests
There were various tests broken because I updated the axios client post
to accept a third argument. This is an optional argument and so
sometimes it is undefined. Jest requires us be explicit in the
.toHaveBeenCalledWith() function, so I needed to add undefined in those
places --- note that there was a lot of updating indents. Most lines in
this change were moved over for formatting, without content changes.

I also updated the mocked response for sending consumer reports, so that
it could run successfully. I updated the response data, and removed the
.toHaveBeenCalledWith() function because it gets called with a brand new
object, and isn't really testable any longer.