Remove lodash pick dependency
While updating dependencies I saw lodash pick is only used in two spots.
Since most lodash functions can be accomplished with pure ES6 syntax I
refactored the code using pick so we can remove lodash pick. That said,
unfortunately, the pick method doesn't transfer super nicely because of
how we're using it, but the code is still clean and readable.

Note: We could have made the code a little cleaner, but since the user
dynamo object is saved in dynamo, and we do not want extra content
saved, we had to be a little more declarative about how we pull out the
user fields.

API-5279