Resolve eslint no-unsafe-call rule failures
This resolves the errors in linting from code not respecting the
no-unsafe-call rule.

As a part of this I had to add an okta-sdk.d.ts file because there are
no type definitions. As a part of this, I moved some of our okta related
definitions into this .d.ts file because they were needed to define the
okta types properly.

API-5355