API-5279: Update dependencies (#114)
* Update all dependencies to wanted versions

I ran npm outdated and npm update to get our packages updated to the
latest 'wanted' versions.

* Remove lodash pick dependency

While updating dependencies I saw lodash pick is only used in two spots.
Since most lodash functions can be accomplished with pure ES6 syntax I
refactored the code using pick so we can remove lodash pick. That said,
unfortunately, the pick method doesn't transfer super nicely because of
how we're using it, but the code is still clean and readable.

Note: We could have made the code a little cleaner, but since the user
dynamo object is saved in dynamo, and we do not want extra content
saved, we had to be a little more declarative about how we pull out the
user fields.

* Update morgan dependency to latest

* Update express dependency to latest

* Update typescript dependency to latest

* Update axios dependency to latest

* Update ts-node dependency to latest

* Update sentry/node to 6.2.0 (latest)

* Update @types/node dependency to 14.14.31 (latest)

* Update jest related dependencies to latest

- jest to 26.6.3
- @types/jest to 26.0.20
- ts-jest to 26.5.1

API-5279

* Update aws-sdk to latest

* Update supertest to 6.1.3 (latest)

* Upgrade eslint and plugins to next major versions

Upgraded:
- eslint to 7.0.0
- @typescript-eslint/eslint-plugin to 3.0.0
- @typescript-eslint/parser to 3.0.0

Note: I added .eslintrc.js to the eslintignore file. With the major
change for parser it began giving an error that eslintrc.js was not
included in the project to be linted.

* Update eslint packages to latest

Updated
- eslint to 7.20.0
- @typescript-eslint/eslint-plugin to 4.15.2
- @typescript-eslint/parser to 4.15.2

* Remove extra eslint rule

There was an extra rule (that isn't even valid) in the configuration.

* Run npm audit fix for security vulnerabilities

* Resolve various typescript compile errors

When we upgraded typescript this broke various tests. This resolves
those issues affecting the tests.

* Bump aws-sdk, nock, eslint versions

API-5279