Initialize OKTA_ORG to blank string
When configuring Okta, we have a condition that checks all necessary
environment variables are defined. If they are not, it throws an error.
This works, but after the code that throws the error, the linter is not
able to detect that OKTA_ORG must defined in the scenario we are using
it. Because it thinks it might be undefined, we were getting a linting
error when using it in a string template expression.

This was originally resolved by casting it to a string, but as a better
solution, we are initializing it to an empty string. This gives us a
string type and is also falsy, so our condition checking for the
variables to be defined will still function appropriately.

API-5355