Add commit hash and version via baked env variables
This adds code to bake commit hash and version into our built files.
Since we don't use package.json for versioning, this allows us to inject
these values during the build process within our cicd pipelines.

This generated a file, baking in environment variables kind of like how
React does it. React does a 'search and replace' while we are
overwriting a file and exporting an object that contains the values.

We generate a file 'generated/baked-env.ts' that exports an object
containing all environment variables starting in 'NODE_APP_'. This only
reads variables during the build. So, if we set NODE_APP_EXAMPLE=abc
during the build, then the baked-env.ts will contain NODE_APP_EXAMPLE
and it will be equal to abc.

API-5393