Various fixe and remove baked version from backend
Right now we are focusing on adding the commit hash, so I removed the
baked version.

Side effects of this commit are pretty big.

1. I modified the docker-compose.yml so that it uses the base node12
image rather than the image from the Dockerfile. This keeps it from
needing to run `npm install` and from running `npm run build` which
makes the local start up time much much quicker. Since it's using
ts-node and nodemon still it's quick to reload changes.

2. I added a second docker-compose that retains the functionality of the
original. Used in conjunction with the main docker-compose, this
docker-compose.build.yml ignores the node_modules folder, runs npm
install, and will run npm start - which mimics prod much closer. The
README has been updated with instructions for this Dockerfile

3. The functionality for environment variables may have changed some
from the original concept. Updated the README in the generated folder.
Added comments through files explaining why we're pulling in the
NODE_APP_ arguments in specific places. The basic flow is this:
- When using docker-compose we define an environment variable
NODE_APP_COMMIT_HASH
- Docker-compose takes that environment variable and passes it to
`docker build` as a build argument.
- If not using docker-compose, you just skip and pass it to `docker
build` as a build argument - no need for an environment variable in that
case
- The Dockerfile takes that build argument and adds it as an environment
variable to the docker image.
- When docker runs and the build runs, the commit hash is available via
that environment variable.

API-5393