Move 'verson' route to its own handler file
When implementing a route with the commit hash, I originally had it at
the root app route. This isn't available for us when in an environment
because it's not one of the public routes. This commit adds a public
route '/version' and removes the version info from the root path route.

API-5393