Export function getBakedEnv rather than object
This adds an exported function to the baked-env.ts generated file. The
reason is to make it easier to mock with jest in order to test out the
version handler endpoint (routes/Version.ts).

This also adds a test specifically for the baked-env.ts file so that we
can get passable test coverage.

API-5393