Fix compiler warnings with jest the first runtime
When jest runs the first time, it compiles all of the tests and stores
them in a cache. This meant that even though they were occurring it
wasn't immediately obvious (running a second time made it look like the
issues were resolved). The most noticable location for the issue was
within our build pipeline, because it would consistently happen.

This commit resolves those compiler warnings.

API-5731