Remove generated files
This commit removes the generated files for retrieving the server
version. Instead, since it looks like we use an entire docker image as
our build artifact rather than just the dist/ folder, this commit
relies on the environment variable COMMIT_HASH to present the version.

The Dockerfile uses the ENV command, which permanently sets an
environment variable in the image (this can be overridden if desired).
We were already doing this actually in order to get the environment
variable in the image during the build process - however, it persists to
runtime as well.

The current flow for presenting an environment variable is
1. For local development, add to .env.local or simply set the
environment variable COMMIT_HASH
2. For docker builds, pass COMMIT_HASH as a build argument. The build
argument is passed to the ENV command in the Dockerfile, permanently
setting that environment variable in the image. Again, it can be
overridden if desired, but if it is not, the running docker container
will have the value that was originally passed in as the build argument

API-5393