API 5731: dependency updates (#134)
* Update out of date dependencies

@sentry/node to 6.2.2
aws-sdk to 2.865.0
@types/node t o 14.14.35
@typescript-eslint/eslint-plugin to 4.18.0
@typescript-eslint/parser to 4.18.0
eslint to 7.22.0
ts-jest to 26.5.4

API-5731

* Fix compiler warnings with jest the first runtime

When jest runs the first time, it compiles all of the tests and stores
them in a cache. This meant that even though they were occurring it
wasn't immediately obvious (running a second time made it look like the
issues were resolved). The most noticable location for the issue was
within our build pipeline, because it would consistently happen.

This commit resolves those compiler warnings.

API-5731

* Revert change to resolve test compiler warnings

I had a change to resolve some compiler warnings that were occuring when
test cases were run. However that change creates a linting error.
Resolving the linting error in the way I attempted, causes other
compiler warnings.

We'll create a separate ticket to track this.

API-5731

* Force react-dev-utils to 11.0.4

The node-mock-server has react-dev-utils as a dependency. That version
of react-dev-utils has a security vulnerability. This will force that
dependency version to resolve the security vulnerability.

API-5731

* Update mock server dependencies

This also adds a resolution for react-dev-utils to set it to 11.0.4, in
order to resolve a security vulnerability.

API-5731

* Remove unneeeded resolution from main package.json

I added a resolution for react-dev-utils to the main package.json. It
turns out there's a package json under the /dev/mocks folder that needed
the resolution. Since the resolution is added there, it's not needed in
the main package.json.

API-5731

* Upgrade dependencies

aws-sdk to 2.867.0
@types/jest to 26.0.21

API-5731