Install node_modules in local dev container
I removed the node_modules installation in the local development
container when removing an unnecessary call to `npm run build`. After
discussion with other devs, we've decided to readd `npm install` to the
default local dev container.

To do this, I added a new image to the Dockerfile, that contains the npm
install, but does not contain the call to `npm run build`. Updated cicd
pipelines, and updated the docker-compose files.

API-5393