Overhaul dockercompose in local development
When using the docker compose to run the backend in local environments,
developers had to wait on a full build, even though the container runs
with ts-node by default. ts-node does not depend on the output of the
build, making the build a time waste.

This adds a third stage to the Dockerfile, enabling us to cutout the
build from the default docker-compose. Instructions were added to the
README.