Add endpoint to respond with commit hash (#138)
This adds the endpoint /internal/developer-portal/public/version that
responds with JSON including the commithash of the deployed code.

The commit hash is received via an environment variable. For local
development, this can be added to the user's '.env' file. For production
deployments, a build argument is  passed to the docker build command,
which then sets a default environment variable in the image with the
value of the build arg/commit hash. When that docker image is run, it
will have the same value it had when the image was built - without
needing to inject it within the deployment pipeline.

The cicd pipeline has been updated to include the build arguments when
running the `docker build` commands.

API-5393