Add query okta id to queryConsumers script
This adds the ability to query by the okta id in the query consumers script.
This also adds merging logic - which will merge the okta id and api fields of
the returned results from Dynamo. Note: This is only merging the returned
dynamo fields, so it won't necessarily have all apis associated with the email
address - because all records with that email address might not be returned.

API-6850