All functionality and tests complete
Emails are being sent as expected to the statusUpdateEmails provided in
req.body and to the SUPPORT_EMAIL defined in environment variables.
Img sources for the production consumer template are the public object URLs and are fine
to be hard coded in the HTML element. An integration test for the route
has been added with this commit as well. All tests are passing and
running `npm run lint` returns no errors.

Signed-off-by: Braden Shipley <bradenrshipley@gmail.com>